requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
grpcio==1.51.1
idna==3.4
iterators==0.0.2
numpy==1.24.2
Pillow==9.4.0
protobuf==3.20.3
PyYAML==6.0
requests==2.28.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.14
